Added gdt; Fixed gitignore (remove previously tracked files)

+++ include/globals.h
@@ -6,8 +6,8 @@
 #ifndef GLOBALS_H
 #define GLOBALS_H
 
-extern uint32_t g_tty_cursor_x;
-extern uint32_t g_tty_cursor_y;
+extern uint64_t g_tty_cursor_x;
+extern uint64_t g_tty_cursor_y;
 extern color_t g_tty_current_color;
 
 extern struct limine_framebuffer *g_vga_active_framebuffer;
