+++ include/graphics/vga.h
@@ -3,7 +3,7 @@
 #ifndef VGA_H
 #define VGA_H
 
-typedef uint32_t color_t;
+typedef uint64_t color_t;
 
 void graphics_init(struct limine_framebuffer *framebuffer);
 void putpixel(uint32_t x, uint32_t y, color_t color);
