+++ src/globals.c
@@ -1,7 +1,7 @@
 #include "globals.h"
 
-uint32_t g_tty_cursor_x = 0;
-uint32_t g_tty_cursor_y = 0;
+uint64_t g_tty_cursor_x = 0;
+uint64_t g_tty_cursor_y = 0;
 color_t g_tty_current_color = PROTO_WHITE;
 
 struct limine_framebuffer *g_vga_active_framebuffer;
